// src/utils/enums.ts
export enum NextEnvironment {
    DEVELOPMENT = 'development',
    STAGING = 'staging',
    PRODUCTION = 'production',
}

export enum Farm {
    UP = 'up',
    UP_ETH = 'up-eth',
    WETH = 'weth',
    WBTC = 'wbtc',
}

export enum FarmToken {
    UP = '$UP',
    UP_ETH = '$UP-ETH',
    WETH = 'WETH',
    WBTC = 'WBTC',
}